Refactor Types and states, add sketch for RequestDevice

+++ cpp/source/event/bridge/DeviceList.cpp
@@ -6,7 +6,7 @@ DeviceList::DeviceList(const skylync::DeviceList& _deviceList):
     Event(Event::DEVICE_LIST)
 {
     deviceList.resize(_deviceList.devices_size());
-    for (size_t i = 0; i < _deviceList.devices_size(); ++i)
+    for (int i = 0; i < _deviceList.devices_size(); ++i)
     {
         const skylync::DeviceId& device = _deviceList.devices(i);
         deviceList.push_back(RefName(device.ref(), device.name()));
